Move some of the searching logic for WritableStorageBackend to an isolated class.
